+++ bindings/nostr-sdk-ffi/src/protocol/types/contact.rs
@@ -2,51 +2,33 @@
 // Copyright (c) 2023-2024 Rust Nostr Developers
 // Distributed under the MIT software license
 
-use std::ops::Deref;
 use std::sync::Arc;
 
 use nostr::RelayUrl;
-use uniffi::Object;
+use uniffi::Record;
 
-use crate::error::Result;
+use crate::error::{NostrSdkError, Result};
 use crate::protocol::key::PublicKey;
 
-#[derive(Debug, PartialEq, Eq, Hash, Object)]
-#[uniffi::export(Debug, Eq, Hash)]
+#[derive(Record)]
 pub struct Contact {
-    inner: nostr::Contact,
+    pub public_key: Arc<PublicKey>,
+    pub relay_url: Option<String>,
+    pub alias: Option<String>,
 }
 
-impl Deref for Contact {
-    type Target = nostr::Contact;
+impl TryFrom<Contact> for nostr::Contact {
+    type Error = NostrSdkError;
 
-    fn deref(&self) -> &Self::Target {
-        &self.inner
-    }
-}
-
-#[uniffi::export]
-impl Contact {
-    #[uniffi::constructor(default(relay_url = None, alias = None))]
-    pub fn new(pk: &PublicKey, relay_url: Option<String>, alias: Option<String>) -> Result<Self> {
-        let relay_url = match relay_url {
+    fn try_from(contact: Contact) -> Result<Self, Self::Error> {
+        let relay_url = match contact.relay_url {
             Some(url) => Some(RelayUrl::parse(&url)?),
             None => None,
         };
-        Ok(Self {
-            inner: nostr::Contact::new(**pk, relay_url, alias),
-        })
-    }
-
-    pub fn alias(&self) -> Option<String> {
-        self.inner.alias.clone()
-    }
-
-    pub fn public_key(&self) -> Arc<PublicKey> {
-        Arc::new(self.inner.public_key.into())
-    }
-
-    pub fn relay_url(&self) -> Option<String> {
-        self.inner.relay_url.clone().map(|u| u.to_string())
+        Ok(nostr::Contact::new(
+            **contact.public_key,
+            relay_url,
+            contact.alias,
+        ))
     }
 }